refactor(cli): extract geometry validation and warning modules
Move the remaining two logical groups out of main.rs:

geometry_validation.rs:
- sinusoidal_a4_topology_supported (wire-chain collinearity gate)
- segment_intersection_error (crossing-wire fail-fast)
- source_risk_geometry_error (tiny-segment EX fail-fast)
- buried_wire_geometry_error (PH2-CHK-002 guardrail)
- private helpers: points_close, segments_share_endpoint,
  segment_closest_distance_and_params, dot3, find_or_insert_node

warnings.rs:
- warn_execution_mode_fallback
- warn_pulse_mode_experimental
- warn_deferred_ground_model
- warn_ge_ground_reflection_flag
- warn_ex_type3_normalization_semantics
- warn_ex_type1/2/4/5_portability_semantics
- warn_pt_card_deferred_support
- warn_nt_card_deferred_support

main.rs now contains only: enums/constants, impl blocks,
BenchRecord struct, bench-emit helpers, and fn main().
